==> code/01_local_CLIF_prone_incidence_analysis_wCOVID.Rmd <==
---
title: "CLIF_prone_incidence_analysis"
author: "Anna Barker and Chad Hochberg"
date: "`r Sys.Date()`"
output: html_document
---
#This File is Run AFTER the Cohort Identification R Markdown ('cohort_identification_prone_incidence.Rmd')
#Specify File Paths, Project Path and Site - SITE SPECIFIC
#Load the Analytic File
```{r}

site_time_zone <-"America/New_York" #Add your local time zone prior to UTC conversion, options for this Analysis are EST = 'America/New_York', CST = 'America/Chicago', PST = 'America/Los_Angeles' 
tables_location <-"Z:/barker/clif_3_22/CLIF-2.0.0-dttm-update-R"
project_location <-"Z:/barker/clif_3_22/" #Add local path to project location, where results will output
site <-"U Michigan" #Add local site name
file_type <-"parquet" #Options are "parquet" or "csv"

check_timezone <- function(site_timezone, intended_timezone = Sys.timezone()) {
  if (site_timezone != intended_timezone) {
    cat("Warning: You specified", site_timezone, "but your system time zone is", intended_timezone, "\n")
    response <- readline(prompt = "Was that intentional? (yes/no): ")
    
    if (tolower(response) %in% c("no", "n")) {
      new_timezone <- readline(prompt = "Please enter the correct timezone: ")
      cat("Timezone updated to", new_timezone, "\n")
      return(new_timezone)
    } else {
      cat("Proceeding with the originally specified timezone:", site_timezone, "\n")
      return(site_timezone)
    }
    
  } else {
    cat("Timezone is set to", site_timezone, "\n")
    return(site_timezone)
  }
}


#Time Zone Check
result <- check_timezone(site_time_zone)
```


```{r Load Needed Libraries, include=FALSE}
packages <- c("lubridate", 
              "tidyverse", 
              "dplyr",
              "tableone", 
              "readxl", 
              "arrow", 
              "collapse", 
              "data.table",
              "viridis",
              "sandwich", #Cluster Robust Standard Errors
              "lmtest",  #Allows Calculation of Cluster Robust Standard Errors
              "marginaleffects") #Predictions and Average Slopes for Average Marginal Effects

install_if_missing <- function(package) {
  if (!require(package, character.only = TRUE)) {
    install.packages(package, dependencies = TRUE)
    library(package, character.only = TRUE)
  }
}

sapply(packages, install_if_missing)
rm(packages, install_if_missing)

#Use Dplyr select as default
select <- dplyr::select

#Create Sub Folders for Graphs within Project Output Folder
# Check if the output directory exists; if not, create it
if (!dir.exists(paste0(project_location, "/project_output/graphs"))) {
  dir.create(paste0(project_location, "/project_output/graphs"))
}

```

```{r Function to Convert Datetimes to Local Time Zone}
#Write Function to Convert Datetime Columns to 'UTC'
fn.timezone_tolocal <- function(df) {
  # Get the dataset schema
  sch <- df$schema
  #Which Fields are Datetime Types?
  datetime_fields <- keep(sch$fields, function(field) {
    grepl("^timestamp", field$type$ToString())
  })
  # Extract the names of these datetime columns
  datetime_cols <- map_chr(datetime_fields, "name")
  
  #Now Convert DateTime Columns to Local Timezone
  df <- df %>%
    mutate(across(all_of(datetime_cols), ~ cast(.x, arrow::timestamp("us", site_time_zone)))) |>
    compute()
}
```



```{r Load Site-specific analytic data set and Convert to Local Time Zone}
#NOTE: This Table is Created by the 
prone_analytic_df <- open_dataset(paste0(project_location, '/project_tables/prone_analytic_data.parquet')) 

#Apply Function to Convert to Local Timezone
prone_analytic_df <- fn.timezone_tolocal(prone_analytic_df)
  
#Bring Into R Environment as Dataframe
prone_analytic_df <- prone_analytic_df |> collect()
```


#Additional Cleaning of Analytic Dataset
#Includes Finalzing Covariates
```{r Define Study Months/Quarters}
#Further Define sex_category as female 1, male/other/unknown 0
prone_analytic_df <- prone_analytic_df |>
  mutate(female=fifelse(tolower(sex_category)=='female', 1, 0)) |>
  #Define Calendar Time for Time of Enrollment
  mutate(year=year(t_enrollment)) |>
  mutate(calendar_month=month(t_enrollment)) |>
  arrange(year, calendar_month)

#Report Out Minimum and Maximum Times
cat('At', site, 'the first eligible patient was enrolled on', as.character(first(prone_analytic_df$t_enrollment)), 
    '\n and the last on', as.character(last(prone_analytic_df$t_enrollment)))

#Create a Study Month Data Frame to Merge to Create Study Month; Can Adjust ym() to accomadte desired period; for this analysis 2018 through 2024
study_month <- data.frame(start_date=seq(ym("201801"), ym("202412"), by= "months")) |>
  mutate(calendar_month=month(start_date)) |>
  mutate(year=year(start_date)) |>
  arrange(year, calendar_month) |>
  mutate(study_month=seq(1:n())) |>
  select(-start_date) |>
  mutate(study_quarter=ceiling(study_month/3))

#Now Merge to Prone Analytic_DF
prone_analytic_df <- prone_analytic_df |>
  left_join(study_month) 
 
#Define the Pre-Specified Study Periods
#Jan 2018-Feb 2020 'Pre-COVID', March 2020-February 2022 'COVID', March 2022-December 2023 'Post_COVID'
#So that All Sites Can Participate Will use the COVID Period as The Reference
prone_analytic_df <- prone_analytic_df |>
  mutate(study_period=fcase(
    study_month>=1 & study_month<27, 'Pre-COVID',
    study_month>=27 & study_month<51, 'COVID',
    study_month>=51, 'Post-COVID'
  )) |>
  mutate(study_period_cat=factor(study_period,
                               levels = c("COVID", "Pre-COVID", "Post-COVID"),
                               labels = c(1, 2, 3)))

#In this script we keep track of the minimum and maximum study month both for graphs and for the modelling
min_month <- min(prone_analytic_df$study_month)
max_month <- max(prone_analytic_df$study_month)
min_quarter <- min(prone_analytic_df$study_quarter)
max_quarter <- max(prone_analytic_df$study_quarter)

#Create 'site_type' vector that shows Which study periods are Included
site_type <- prone_analytic_df |>
  distinct(study_period) 
site_type <- paste(site_type$study_period, collapse = ",")
#Possible Combinations are "Pre-COVID,COVID,Post-COVID"|"COVID,Post-COVID"|"Pre-COVID,"COVID"

#Keep Track of the Number of Hospitals
n_hospitals <- length(unique(prone_analytic_df$hospital_id[!is.na(prone_analytic_df$hospital_id)]))

#Create a Few More Variables
prone_analytic_df <- prone_analytic_df |>
  mutate(admit_to_enrolled=as.duration(t_enrollment-final_admission_dttm)/dhours(1)) |>
  mutate(ett_to_enrolled=as.duration(t_enrollment-vent_episode_start)/dhours(1)) |>
  mutate(severe_ards=fifelse(min_pf_ratio<100, 1, 0))

#Create a 'Unknown' Category for SARS COV2 - THis is if we can't confirm a positive or negative SARS-Cov2 test in the COVID or Post-COVID era
prone_analytic_df <- prone_analytic_df |>
  mutate(sars_cov2_positive=fcase(
    sars_cov2_positive==1, 'Positive', 
    sars_cov2_positive==0, 'Negative',
    study_period_cat %in% c(1, 3) & is.na(sars_cov2_positive), 'Unknown',
    study_period_cat %in% c(2), 'Pre-COVID'
  )) |>
  #Create Period-Sars-COV2 Status Categorical Variables - For Categorical Variable COVID-Negative Could be Reference
  mutate(period_sarscov2=fcase(
    study_period_cat==2, 'Pre-COVID', 
    study_period_cat==1 & sars_cov2_positive=='Positive', 'COVID_SarsCov2Pos',
    study_period_cat==1 & sars_cov2_positive %in% c('Negative', 'Unknown'), 'COVID_SarsCov2Neg-Unk',
    study_period_cat==3 & sars_cov2_positive=='Positive','Post-COVID_SarsCov2Pos',
    study_period_cat==3 & sars_cov2_positive %in% c('Negative', 'Unknown'), 'Post-COVID_SarsCov2Neg-Unk'
  )) |>
  #Given the POST-covid Sars_cov2 + is a small population will collapse that into post-COVID for some analysis, COVID-Negative is Reference
  mutate(period_sarscov2_post=fcase(
    study_period_cat==2, 'Pre-COVID', 
    study_period_cat==1 & sars_cov2_positive=='Positive', 'COVID_SarsCov2Pos',
    study_period_cat==1 & sars_cov2_positive %in% c('Negative', 'Unknown'), 'COVID_SarsCov2Neg-Unk',
    study_period_cat==3,'Post-COVID'
  ))
```


```{r Describe The Number of Admissions Per Month, Admissions Per Period}
#Create a Table of N of Patients and Number Proned Per Study Month
prone_per_month <- prone_analytic_df |> 
  group_by(study_month) |>
  summarise(
    'site' = site,
    'n_hospitals' = n_hospitals,
    'year' = mean(year),
    'calendar_month' = mean(calendar_month),
    'study_period' = max(study_period),
    'n_patients' = n(),
    'n_proned_12_hr' = sum(prone_12hour_outcome),
    'proned_12_hr_percent' = mean(prone_12hour_outcome),
    'standard_error_12' = sqrt(mean(prone_12hour_outcome) * (1 - mean(prone_12hour_outcome)) / n()),
    'n_proned_24_hr' = sum(prone_24hour_outcome),
    'proned_24_hr_percent' = mean(prone_24hour_outcome),
    'standard_error_24' = sqrt(mean(prone_24hour_outcome) * (1 - mean(prone_24hour_outcome)) / n()),
    'n_proned_72_hr' = sum(prone_72hour_outcome),
    'proned_72_hr_percent' = mean(prone_72hour_outcome),
    'standard_error_72' = sqrt(mean(prone_72hour_outcome) * (1 - mean(prone_72hour_outcome)) / n()),
    'n_proned_all' = sum(proned),
    'proned_percent_all' = mean(proned),
    'standard_error_all' = sqrt(mean(proned) * (1 - mean(proned)) / n())
  ) |>
  relocate(site, n_hospitals, .before = study_month)
#Save Locally At Each Site but Don't Export as May have 1 Patient Per Month
write_csv(prone_per_month, paste0(project_location, '/project_tables/', site,'_prone_per_month.csv'))

#Create a Table of N of Patients and Number Proned Per Study Month Per Hospital
prone_per_month_by_hospital <- prone_analytic_df |> 
  group_by(study_month, hospital_id) |>
  summarise(
    'site' = site,
    'n_hospitals' = n_hospitals,
    'year' = mean(year),
    'calendar_month' = mean(calendar_month),
    'study_period' = max(study_period),
    'n_patients' = n(),
    'n_proned_12_hr' = sum(prone_12hour_outcome),
    'proned_12_hr_percent' = mean(prone_12hour_outcome),
    'standard_error_12' = sqrt(mean(prone_12hour_outcome) * (1 - mean(prone_12hour_outcome)) / n()),
    'n_proned_24_hr' = sum(prone_24hour_outcome),
    'proned_24_hr_percent' = mean(prone_24hour_outcome),
    'standard_error_24' = sqrt(mean(prone_24hour_outcome) * (1 - mean(prone_24hour_outcome)) / n()),
    'n_proned_72_hr' = sum(prone_72hour_outcome),
    'proned_72_hr_percent' = mean(prone_72hour_outcome),
    'standard_error_72' = sqrt(mean(prone_72hour_outcome) * (1 - mean(prone_72hour_outcome)) / n()),
    'n_proned_all' = sum(proned),
    'proned_percent_all' = mean(proned),
    'standard_error_all' = sqrt(mean(proned) * (1 - mean(proned)) / n())
  ) |>
  relocate(site, n_hospitals, .before = study_month)
write_csv(prone_per_month_by_hospital, paste0(project_location, '/project_tables/', site,'_prone_per_month_hospital.csv'))


#Repeat Per Study Period
prone_per_period <- prone_analytic_df |> 
  group_by(study_period) |>
  summarise(
    'site' = site,
    'n_hospitals' = n_hospitals,
    'n_patients' = n(),
    'n_proned_12_hr' = sum(prone_12hour_outcome),
    'proned_12_hr_percent' = mean(prone_12hour_outcome),
    'standard_error_12' = sqrt(mean(prone_12hour_outcome) * (1 - mean(prone_12hour_outcome)) / n()),
    'n_proned_24_hr' = sum(prone_24hour_outcome),
    'proned_24_hr_percent' = mean(prone_24hour_outcome),
    'standard_error_24' = sqrt(mean(prone_24hour_outcome) * (1 - mean(prone_24hour_outcome)) / n()),
    'n_proned_72_hr' = sum(prone_72hour_outcome),
    'proned_72_hr_percent' = mean(prone_72hour_outcome),
    'standard_error_72' = sqrt(mean(prone_72hour_outcome) * (1 - mean(prone_72hour_outcome)) / n()),
    'n_proned_all' = sum(proned),
    'proned_percent_all' = mean(proned),
    'standard_error_all' = sqrt(mean(proned) * (1 - mean(proned)) / n())
  ) |>
  relocate(site, n_hospitals, .before = study_period)
write_csv(prone_per_period, paste0(project_location, '/project_output/', site,'_prone_per_period.csv'))

#Repeat Per Study Quarter
prone_per_quarter <- prone_analytic_df |> 
  group_by(study_quarter) |>
  summarise(
    'site' = site,
    'n_hospitals' = n_hospitals,
    'study_period' = max(study_period),
    'n_patients' = n(),
    'sars_cov2_positive' = sum(sars_cov2_positive=='Positive', na.rm=T),
    'n_proned_12_hr' = sum(prone_12hour_outcome),
    'proned_12_hr_percent' = mean(prone_12hour_outcome),
    'standard_error_12' = sqrt(mean(prone_12hour_outcome) * (1 - mean(prone_12hour_outcome)) / n()),
    'n_proned_24_hr' = sum(prone_24hour_outcome),
    'proned_24_hr_percent' = mean(prone_24hour_outcome),
    'standard_error_24' = sqrt(mean(prone_24hour_outcome) * (1 - mean(prone_24hour_outcome)) / n()),
    'n_proned_72_hr' = sum(prone_72hour_outcome),
    'proned_72_hr_percent' = mean(prone_72hour_outcome),
    'standard_error_72' = sqrt(mean(prone_72hour_outcome) * (1 - mean(prone_72hour_outcome)) / n()),
    'n_proned_all' = sum(proned),
    'proned_percent_all' = mean(proned),
    'standard_error_all' = sqrt(mean(proned) * (1 - mean(proned)) / n())
      ) |>
  relocate(site, n_hospitals, .before = study_quarter)
write_csv(prone_per_quarter, paste0(project_location, '/project_output/', site,'_prone_per_quarter.csv'))

#Repeat Per Study Period and sars_cov2_status
prone_per_quarter_sarscov2 <- prone_analytic_df |> 
  mutate(sars_cov2_positive=fcase(
    sars_cov2_positive=='Positive', 'COVID', 
    default = 'Not COVID'
  )) |>
  group_by(study_quarter, sars_cov2_positive) |>
  summarise(
    'site' = site,
    'n_hospitals' = n_hospitals,
    'study_period' = max(study_period),
    'n_patients' = n(),
    'n_proned_12_hr' = sum(prone_12hour_outcome),
    'proned_12_hr_percent' = mean(prone_12hour_outcome),
    'standard_error_12' = sqrt(mean(prone_12hour_outcome) * (1 - mean(prone_12hour_outcome)) / n()),
    'n_proned_24_hr' = sum(prone_24hour_outcome),
    'proned_24_hr_percent' = mean(prone_24hour_outcome),
    'standard_error_24' = sqrt(mean(prone_24hour_outcome) * (1 - mean(prone_24hour_outcome)) / n()),
    'n_proned_72_hr' = sum(prone_72hour_outcome),
    'proned_72_hr_percent' = mean(prone_72hour_outcome),
    'standard_error_72' = sqrt(mean(prone_72hour_outcome) * (1 - mean(prone_72hour_outcome)) / n()),
    'n_proned_all' = sum(proned),
    'proned_percent_all' = mean(proned),
    'standard_error_all' = sqrt(mean(proned) * (1 - mean(proned)) / n())
      ) |>
  relocate(site, n_hospitals, .before = sars_cov2_positive)
write_csv(prone_per_quarter_sarscov2, paste0(project_location, '/project_output/', site,'_prone_per_quarter-sarscoV2_status.csv'))

#Create a Histogram of Admissions Per Study Month by Hospital
#Create Month Labels - Will Use Study Specific Min and Max Months to Select The Right Grouping
month_labels <- c(
'Jan18', 'Feb18', 'Mar18','Apr18', 
'May18', 'Jun18', 'Jul18', 'Aug18',
'Sep18', 'Oct18', 'Nov18', 'Dec18',
'Jan19', 'Feb19', 'Mar19','Apr19', 
'May19', 'Jun19', 'Jul19', 'Aug19',
'Sep19', 'Oct19', 'Nov19', 'Dec19',
'Jan20', 'Feb20', 'Mar20','Apr20', 
'May20', 'Jun20', 'Jul20', 'Aug20',
'Sep20', 'Oct20', 'Nov20', 'Dec20',
'Jan21', 'Feb21', 'Mar21','Apr21', 
'May21', 'Jun21', 'Jul21', 'Aug21',
'Sep21', 'Oct21', 'Nov21', 'Dec21',
'Jan22', 'Feb22', 'Mar22','Apr22', 
'May22', 'Jun22', 'Jul22', 'Aug22',
'Sep22', 'Oct22', 'Nov22', 'Dec22',
'Jan23', 'Feb23', 'Mar23', 'Apr23', 
'May23', 'Jun23', 'Jul23', 'Aug23', 
'Sep23', 'Oct23', 'Nov23', 'Dec23',
'Jan24', 'Feb24', 'Mar24', 'Apr24', 
'May24', 'Jun24', 'Jul24', 'Aug24', 
'Sep24', 'Oct24', 'Nov24', 'Dec24')

#Select the Relevant Months from This Vector Using the min_month, max_month
site_month_labels <- month_labels[min_month:max_month]
max_monthly_n <- max(prone_per_month$n_patients)

#Plot of Enrollments Per Month
ggplot(prone_analytic_df, aes(x=study_month, fill=hospital_id)) +
  geom_bar(position = 'stack') +
  theme_classic() +
  scale_x_continuous(breaks=seq(min_month, max_month, by=1), limits = c(min_month, max_month),
                     labels = site_month_labels,
                     name = 'Month-Year') +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1, size = 4.75)) +
  scale_y_continuous(name = 'Number of Admissions') +
  scale_fill_viridis_d()
ggsave(paste0(site, '_enrollment_per_month.pdf'),
       path = paste0(project_location, '/project_output/graphs/'))

#Now Plot by sars_cov2_status
ggplot(prone_analytic_df, aes(x=study_month, fill=factor(sars_cov2_positive))) +
  geom_bar(position = 'stack') +
  theme_classic() +
  scale_x_continuous(breaks=seq(min_month, max_month, by=1), limits = c(min_month, max_month),
                     labels = site_month_labels,
                     name = 'Month-Year') +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1, size = 4.75)) +
  scale_y_continuous(name = 'Number of Admissions') +
  scale_fill_viridis_d()
ggsave('admits_by_month_by_covid.pdf',
       path = 'project_output/graphs/')

```

```{r Table 1 by Period and By Hospital}
#Variables to Table 1
to_tab <- c(
  'age_at_admission',
  'female',
  'race_ethnicity',
  'bmi',
  'hospital_id',
  'year',
  'study_period',
  'period_sarscov2',
  'eligible_by_proseva',
  'eligible_by_prone',
  'admit_to_enrolled', 
  'ett_to_enrolled',
  'or_before_enrollment',
  'min_pf_ratio',
  'severe_ards',
  'first_proseva_mode',
  'first_proseva_peep',
  'first_proseva_fio2',
  'sofa_score',
  'nee_pressor_dose'
)

#Factor Variables
factors_tab <- c(
  'female',
  'race_ethnicity',
  'hospital_id',
  'year',
  'study_period',
  'period_sarscov2',
  'or_before_enrollment',
  'first_proseva_mode',
  'nee_pressor_dose',
  'severe_ards',
  'eligible_by_proseva',
  'eligible_by_prone'
)

#First Create Table with Ranges and Additional Statistics Including Numbers of Missing and Min/Max
tab1 <- CreateTableOne(vars = to_tab, data=prone_analytic_df, factorVars = factors_tab)
summary(tab1) 

#Now Create a Traditional Table 1, One Stratified by Period and One Stratified by Hospital
tab1 <- CreateTableOne(vars = to_tab, strata="study_period", data=prone_analytic_df, factorVars = factors_tab, addOverall = TRUE) 
print(tab1)
tab1_excel <- as.data.frame(print(tab1, printToggle = FALSE, missing = TRUE)) |>
  mutate(site = site) |>
  relocate(site, .before = Overall) |>
  rownames_to_column('characteristic') |>
  rename(percent_missing = Missing) |>
  select(-test)
write.csv(tab1_excel, file=paste0(project_location, '/project_output/', site, '_Table1_by_Period.csv'))
rm(tab1_excel)

#Now Table 1, One Stratified by Hospital
if (n_hospitals>1) {
tab1 <- CreateTableOne(vars = to_tab, strata="hospital_id", data=prone_analytic_df, factorVars = factors_tab, addOverall = TRUE) 
print(tab1)
tab1_excel <- as.data.frame(print(tab1, printToggle = FALSE, missing = TRUE)) |>
  mutate(site = site) |>
  relocate(site, .before = Overall) |>
  rownames_to_column('characteristic') |>
  rename(percent_missing = Missing) |>
  select(-test)
} else {
tab1 <- CreateTableOne(vars = to_tab, data=prone_analytic_df, factorVars = factors_tab, addOverall = TRUE) 
print(tab1)
tab1_excel <- as.data.frame(print(tab1, printToggle = FALSE, missing = TRUE)) |>
  mutate(site = site) |>
  relocate(site, .before = Overall) |>
  rownames_to_column('characteristic') |>
  rename(percent_missing = Missing)
}
write.csv(tab1_excel, file=paste0(project_location, '/project_output/', site, '_Table1_by_Hospital.csv'))
rm(tab1_excel)


#Now Create a Traditional Table 1, Stratified by Period and SARS-Cov2 status within COVID
tab1 <- CreateTableOne(vars = to_tab, strata="period_sarscov2_post", data=prone_analytic_df, factorVars = factors_tab, addOverall = TRUE) 
print(tab1)
tab1_excel <- as.data.frame(print(tab1, printToggle = FALSE, missing = TRUE)) |>
  mutate(site = site) |>
  relocate(site, .before = Overall) |>
  rownames_to_column('characteristic') |>
  rename(percent_missing = Missing) |>
  select(-test)
write.csv(tab1_excel, file=paste0(project_location, '/project_output/', site, '_Table1_by_Period_Post.csv'))
rm(tab1_excel)
```

```{r Visualize Proning Over Time}
#Graph the Proning Rate by Study Quarter with Error Bars for Standard Error
#Create Labels for Study Quarter
quarter_labels <- c(
'2018:1-3', '2018:4-6', '2018:7-9', '2018:10-12',
'2019:1-3', '2019:4-6', '2019:7-9', '2019:10-12',
'2020:1-3', '2020:4-6', '2020:7-9', '2020:10-12',
'2021:1-3', '2021:4-6', '2021:7-9', '2021:10-12',
'2022:1-3', '2022:4-6', '2022:7-9', '2022:10-12',
'2023:1-3', '2023:4-6', '2023:7-9', '2023:10-12',
'2024:1-3', '2024:4-6', '2024:7-9', '2024:10-12')
study_quarter_labels <- quarter_labels[min_quarter:max_quarter]

base_plot <- ggplot(prone_per_quarter, aes(x=study_quarter, y=proned_12_hr_percent)) +
  geom_errorbar(aes(ymin = proned_12_hr_percent - standard_error_12, ymax = proned_12_hr_percent + 
                      standard_error_12), width = 0.1, alpha = 0.5, color = 'grey') +
  geom_point() +
  geom_line() +
  scale_y_continuous(breaks=seq(0,1, by=0.1), 
                     labels = scales::percent, 
                     limits = c(0,1.0),
                     name = 'Proned within 12 Hours of Enrollment') +
  scale_x_continuous(breaks=seq(min_quarter, max_quarter, by=1),
                     labels = study_quarter_labels,
                     limits = c(min_quarter, max_quarter),
                     name = 'Year: Months') +
  theme_classic() + # theme_classic call
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) # custom x-axis text

#Now Add Annotations and Verticle Lines to Plot Based on What Data a Site Has Contributed
if (site_type == "Pre-COVID,COVID,Post-COVID") { #If Site Has Full Study Period (2018-2023)
  extended <- base_plot +
    geom_vline(xintercept = 9.6667, linetype = 2) +
    annotate("text", x=5.667, y=0.95, label = 'Pre COVID-19') +
    geom_vline(xintercept = 17.6667, linetype = 2) +
    annotate("text", x=13.667, y=0.95, label = 'COVID-19') +
    annotate("text", x=21.667, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "COVID,Post-COVID") {
  extended <- base_plot + 
    geom_vline(xintercept = 17.6667, linetype = 2) +
    annotate("text", x=13.667, y=0.95, label = 'COVID-19') +
    annotate("text", x=21.667, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "Pre-COVID,COVID") {
  extended <- base_plot + 
    geom_vline(xintercept = 9.6667, linetype = 2) +
    annotate("text", x=5.667, y=0.95, label = 'Pre COVID-19') +
    annotate("text", x=13.667, y=0.95, label = 'COVID-19') 
} else {
  extended <- base_plot
}
# Display the extended plot
extended
ggsave(paste0(site, '_proning_by_quarter_12.pdf'),
       path = paste0(project_location, '/project_output/graphs/'))

#Repeat for any proning, within 24h
base_plot_24 <- ggplot(prone_per_quarter, aes(x=study_quarter, y=proned_24_hr_percent)) +
  geom_errorbar(aes(ymin = proned_24_hr_percent - standard_error_24, ymax = proned_24_hr_percent + 
                      standard_error_24), width = 0.1, alpha = 0.5, color = 'grey') +
  geom_point() +
  geom_line() +
  scale_y_continuous(breaks=seq(0,1, by=0.1), 
                     labels = scales::percent, 
                     limits = c(0,1.0),
                     name = 'Proned with 24 hours of Meeting Proseva Criteria') +
  scale_x_continuous(breaks=seq(min_quarter, max_quarter, by=1),
                     labels = study_quarter_labels,
                     limits = c(min_quarter, max_quarter),
                     name = 'Year: Months') +
  theme_classic() + # theme_classic call
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) # custom x-axis text

#Now Add Annotations and Verticle Lines to Plot Based on What Data a Site Has Contributed
if (site_type == "Pre-COVID,COVID,Post-COVID") { #If Site Has Full Study Period (2018-2023)
  extended <- base_plot_24 +
    geom_vline(xintercept = 9.6667, linetype = 2) +
    annotate("text", x=5.667, y=0.95, label = 'Pre COVID-19') +
    geom_vline(xintercept = 17.6667, linetype = 2) +
    annotate("text", x=13.667, y=0.95, label = 'COVID-19') +
    annotate("text", x=21.667, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "COVID,Post-COVID") {
  extended <- base_plot_24 + 
    geom_vline(xintercept = 17.6667, linetype = 2) +
    annotate("text", x=13.667, y=0.95, label = 'COVID-19') +
    annotate("text", x=21.667, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "Pre-COVID,COVID") {
  extended <- base_plot_24 + 
    geom_vline(xintercept = 9.6667, linetype = 2) +
    annotate("text", x=5.667, y=0.95, label = 'Pre COVID-19') +
    annotate("text", x=13.667, y=0.95, label = 'COVID-19') 
} else {
  extended <- base_plot_24
}
# Display the extended plot
extended
ggsave(paste0(site, '_proning_by_quarter_24.pdf'),
       path = paste0(project_location, '/project_output/graphs/'))


#Repeat for any proning, within 72h 
base_plot_72 <- ggplot(prone_per_quarter, aes(x=study_quarter, y=proned_72_hr_percent)) +
  geom_errorbar(aes(ymin = proned_72_hr_percent - standard_error_72, ymax = proned_72_hr_percent + 
                      standard_error_72), width = 0.1, alpha = 0.5, color = 'grey') +
  geom_point() +
  geom_line() +
  scale_y_continuous(breaks=seq(0,1, by=0.1), 
                     labels = scales::percent, 
                     limits = c(0,1.0),
                     name = 'Proned Within 72 Hours of Meeting PROSEVA Criteria') +
  scale_x_continuous(breaks=seq(min_quarter, max_quarter, by=1),
                     labels = study_quarter_labels,
                     limits = c(min_quarter, max_quarter),
                     name = 'Year: Months') +
  theme_classic() + # theme_classic call
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) # custom x-axis text

#Now Add Annotations and Verticle Lines to Plot Based on What Data a Site Has Contributed
if (site_type == "Pre-COVID,COVID,Post-COVID") { #If Site Has Full Study Period (2018-2023)
  extended <- base_plot_72 +
    geom_vline(xintercept = 9.6667, linetype = 2) +
    annotate("text", x=5.667, y=0.95, label = 'Pre COVID-19') +
    geom_vline(xintercept = 17.6667, linetype = 2) +
    annotate("text", x=13.667, y=0.95, label = 'COVID-19') +
    annotate("text", x=21.667, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "COVID,Post-COVID") {
  extended <- base_plot_72 + 
    geom_vline(xintercept = 17.6667, linetype = 2) +
    annotate("text", x=13.667, y=0.95, label = 'COVID-19') +
    annotate("text", x=21.667, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "Pre-COVID,COVID") {
  extended <- base_plot_72 + 
    geom_vline(xintercept = 9.6667, linetype = 2) +
    annotate("text", x=5.667, y=0.95, label = 'Pre COVID-19') +
    annotate("text", x=13.667, y=0.95, label = 'COVID-19') 
} else {
  extended <- base_plot_72
}
# Display the extended plot
extended
ggsave(paste0(site, '_proning_by_quarter_72.pdf'),
       path = paste0(project_location, '/project_output/graphs/'))

#Repeat for any proning, all prones among patients in cohort
base_plot_all <- ggplot(prone_per_quarter, aes(x=study_quarter, y=proned_percent_all)) +
  geom_errorbar(aes(ymin = proned_percent_all - standard_error_all, ymax = proned_percent_all + 
                      standard_error_all), width = 0.1, alpha = 0.5, color = 'grey') +
  geom_point() +
  geom_line() +
  scale_y_continuous(breaks=seq(0,1, by=0.1), 
                     labels = scales::percent, 
                     limits = c(0,1.0),
                     name = 'Proned during Admission') +
  scale_x_continuous(breaks=seq(min_quarter, max_quarter, by=1),
                     labels = study_quarter_labels,
                     limits = c(min_quarter, max_quarter),
                     name = 'Year: Months') +
  theme_classic() + # theme_classic call
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) # custom x-axis text

#Now Add Annotations and Verticle Lines to Plot Based on What Data a Site Has Contributed
if (site_type == "Pre-COVID,COVID,Post-COVID") { #If Site Has Full Study Period (2018-2023)
  extended <- base_plot_all +
    geom_vline(xintercept = 9.6667, linetype = 2) +
    annotate("text", x=5.667, y=0.95, label = 'Pre COVID-19') +
    geom_vline(xintercept = 17.6667, linetype = 2) +
    annotate("text", x=13.667, y=0.95, label = 'COVID-19') +
    annotate("text", x=21.667, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "COVID,Post-COVID") {
  extended <- base_plot_all + 
    geom_vline(xintercept = 17.6667, linetype = 2) +
    annotate("text", x=13.667, y=0.95, label = 'COVID-19') +
    annotate("text", x=21.667, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "Pre-COVID,COVID") {
  extended <- base_plot_all + 
    geom_vline(xintercept = 9.6667, linetype = 2) +
    annotate("text", x=5.667, y=0.95, label = 'Pre COVID-19') +
    annotate("text", x=13.667, y=0.95, label = 'COVID-19') 
} else {
  extended <- base_plot_all
}
# Display the extended plot
extended
ggsave(paste0(site, '_proning_by_quarter_all.pdf'),
       path = paste0(project_location, '/project_output/graphs/'))

#Now Repeat for Month
#Label Every 3rd Month
tool <- seq(min_month, max_month, by=2)
new_month_labels <- site_month_labels[tool]
base_plot <- ggplot(prone_per_month, aes(x=study_month, y=proned_12_hr_percent)) +
  geom_errorbar(aes(ymin = proned_12_hr_percent - standard_error_12, ymax = proned_12_hr_percent + standard_error_12), 
                width = 0.1, alpha = 0.5, color = 'grey') +
  geom_point() +
  geom_line() +
  scale_y_continuous(breaks=seq(0,1, by=0.1), 
                     labels = scales::percent, 
                     limits = c(0,1.0),
                     name = 'Proned within 12 Hours of Enrollment') +
  scale_x_continuous(breaks=seq(min_month, max_month, by=2),
                     labels = new_month_labels,
                     limits = c(min_month, max_month),
                     name = 'Month-Year') +
  theme_classic() + # theme_classic call
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) # custom x-axis text

#Now Add Annotations and Verticle Lines to Plot Based on What Data a Site Has Contributed
if (site_type == "Pre-COVID,COVID,Post-COVID") { #If Site Has Full Study Period (2018-2023)
  extended <- base_plot +
    geom_vline(xintercept = 27, linetype = 2) +
    annotate("text", x=15, y=0.95, label = 'Pre COVID-19') +
    geom_vline(xintercept = 51, linetype = 2) +
    annotate("text", x=39, y=0.95, label = 'COVID-19') +
    annotate("text", x=63, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "COVID,Post-COVID") { #Just COVID and Post COVID
  extended <- base_plot + 
    geom_vline(xintercept = 51, linetype = 2) +
    annotate("text", x=39, y=0.95, label = 'COVID-19') +
    annotate("text", x=63, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "Pre-COVID,COVID") { #Just Pre-COVID and COVID
  extended <- base_plot + 
    geom_vline(xintercept = 27, linetype = 2) +
    annotate("text", x=15, y=0.95, label = 'Pre COVID-19') +
    annotate("text", x=39, y=0.95, label = 'COVID-19')
} else {
  extended <- base_plot
}
# Display the extended plot
extended
ggsave(paste0(site, '_proning_by_month_12.pdf'),
       path = paste0(project_location, '/project_output/graphs/'))

#Repeat for monthly proning, within 24h meeting criteria
base_plot_24 <- ggplot(prone_per_month, aes(x=study_month, y=proned_24_hr_percent)) +
  geom_errorbar(aes(ymin = proned_24_hr_percent - standard_error_24, ymax = proned_24_hr_percent + standard_error_24), 
                width = 0.1, alpha = 0.5, color = 'grey') +
  geom_point() +
  geom_line() +
  scale_y_continuous(breaks=seq(0,1, by=0.1), 
                     labels = scales::percent, 
                     limits = c(0,1.0),
                     name = 'Proned Within 24h of Meeting PROSEVA Criteria') +
  scale_x_continuous(breaks=seq(min_month, max_month, by=2),
                     labels = new_month_labels,
                     limits = c(min_month, max_month),
                     name = 'Month-Year') +
  theme_classic() + # theme_classic call
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) # custom x-axis text

#Now Add Annotations and Verticle Lines to Plot Based on What Data a Site Has Contributed
if (site_type == "Pre-COVID,COVID,Post-COVID") { #If Site Has Full Study Period (2018-2023)
  extended <- base_plot_24 +
    geom_vline(xintercept = 27, linetype = 2) +
    annotate("text", x=15, y=0.95, label = 'Pre COVID-19') +
    geom_vline(xintercept = 51, linetype = 2) +
    annotate("text", x=39, y=0.95, label = 'COVID-19') +
    annotate("text", x=63, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "COVID,Post-COVID") { #Just COVID and Post COVID
  extended <- base_plot_24 + 
    geom_vline(xintercept = 51, linetype = 2) +
    annotate("text", x=39, y=0.95, label = 'COVID-19') +
    annotate("text", x=63, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "Pre-COVID,COVID") { #Just Pre-COVID and COVID
  extended <- base_plot_24 + 
    geom_vline(xintercept = 27, linetype = 2) +
    annotate("text", x=15, y=0.95, label = 'Pre COVID-19') +
    annotate("text", x=39, y=0.95, label = 'COVID-19')
} else {
  extended <- base_plot_24
}
# Display the extended plot
extended
ggsave(paste0(site, '_proning_by_month_24.pdf'),
       path = paste0(project_location, '/project_output/graphs/'))

#Repeat for monthly proning, prone within 72 h
base_plot_72 <- ggplot(prone_per_month, aes(x=study_month, y=proned_72_hr_percent)) +
  geom_errorbar(aes(ymin = proned_72_hr_percent - standard_error_72, ymax = proned_72_hr_percent + standard_error_72), 
                width = 0.1, alpha = 0.5, color = 'grey') +
  geom_point() +
  geom_line() +
  scale_y_continuous(breaks=seq(0,1, by=0.1), 
                     labels = scales::percent, 
                     limits = c(0,1.0),
                     name = 'Proned within 72 hours of meeting PROSEVA criteria') +
  scale_x_continuous(breaks=seq(min_month, max_month, by=2),
                     labels = new_month_labels,
                     limits = c(min_month, max_month),
                     name = 'Month-Year') +
  theme_classic() + # theme_classic call
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) # custom x-axis text

#Now Add Annotations and Verticle Lines to Plot Based on What Data a Site Has Contributed
if (site_type == "Pre-COVID,COVID,Post-COVID") { #If Site Has Full Study Period (2018-2023)
  extended <- base_plot_72 +
    geom_vline(xintercept = 27, linetype = 2) +
    annotate("text", x=15, y=0.95, label = 'Pre COVID-19') +
    geom_vline(xintercept = 51, linetype = 2) +
    annotate("text", x=39, y=0.95, label = 'COVID-19') +
    annotate("text", x=63, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "COVID,Post-COVID") { #Just COVID and Post COVID
  extended <- base_plot_72 + 
    geom_vline(xintercept = 51, linetype = 2) +
    annotate("text", x=39, y=0.95, label = 'COVID-19') +
    annotate("text", x=63, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "Pre-COVID,COVID") { #Just Pre-COVID and COVID
  extended <- base_plot_72 + 
    geom_vline(xintercept = 27, linetype = 2) +
    annotate("text", x=15, y=0.95, label = 'Pre COVID-19') +
    annotate("text", x=39, y=0.95, label = 'COVID-19')
} else {
  extended <- base_plot_72
}
# Display the extended plot
extended
ggsave(paste0(site, '_proning_by_month_72.pdf'),
       path = paste0(project_location, '/project_output/graphs/'))

#Repeat for monthly proning, "all_time" proning
base_plot_all <- ggplot(prone_per_month, aes(x=study_month, y=proned_percent_all)) +
  geom_errorbar(aes(ymin = proned_percent_all - standard_error_all, ymax = proned_percent_all + standard_error_all), 
                width = 0.1, alpha = 0.5, color = 'grey') +
  geom_point() +
  geom_line() +
  scale_y_continuous(breaks=seq(0,1, by=0.1), 
                     labels = scales::percent, 
                     limits = c(0,1.0),
                     name = 'Proned during Admission') +
  scale_x_continuous(breaks=seq(min_month, max_month, by=2),
                     labels = new_month_labels,
                     limits = c(min_month, max_month),
                     name = 'Month-Year') +
  theme_classic() + # theme_classic call
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) # custom x-axis text

#Now Add Annotations and Verticle Lines to Plot Based on What Data a Site Has Contributed
if (site_type == "Pre-COVID,COVID,Post-COVID") { #If Site Has Full Study Period (2018-2023)
  extended <- base_plot_all +
    geom_vline(xintercept = 27, linetype = 2) +
    annotate("text", x=15, y=0.95, label = 'Pre COVID-19') +
    geom_vline(xintercept = 51, linetype = 2) +
    annotate("text", x=39, y=0.95, label = 'COVID-19') +
    annotate("text", x=63, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "COVID,Post-COVID") { #Just COVID and Post COVID
  extended <- base_plot_all + 
    geom_vline(xintercept = 51, linetype = 2) +
    annotate("text", x=39, y=0.95, label = 'COVID-19') +
    annotate("text", x=63, y=0.95, label = 'Post COVID-19') 
} else if (site_type == "Pre-COVID,COVID") { #Just Pre-COVID and COVID
  extended <- base_plot_all + 
    geom_vline(xintercept = 27, linetype = 2) +
    annotate("text", x=15, y=0.95, label = 'Pre COVID-19') +
    annotate("text", x=39, y=0.95, label = 'COVID-19')
} else {
  extended <- base_plot_all
}
# Display the extended plot
extended
ggsave(paste0(site, '_proning_by_month_all.pdf'),
       path = paste0(project_location, '/project_output/graphs/'))

##Finally Plot By SARS-COV2 and Period
base_plot_covid <- ggplot(prone_per_quarter_sarscov2, aes(x = study_quarter, y = proned_12_hr_percent)) +
  geom_errorbar(aes(ymin = proned_12_hr_percent - standard_error_12,
                    ymax = proned_12_hr_percent + standard_error_12),
                width = 0.1, alpha = 0.5, color = 'grey') +
  geom_point(aes(color = sars_cov2_positive)) +
  geom_line(aes(color = sars_cov2_positive, group = sars_cov2_positive)) +
  scale_y_continuous(breaks=seq(0,1, by=0.1), 
                     labels = scales::percent, 
                     limits = c(0,1.0),
                     name = 'Proned within 12 Hours of Enrollment') +
  scale_x_continuous(breaks=seq(min_quarter, max_quarter, by=1),
                     labels = study_quarter_labels,
                     limits = c(min_quarter, max_quarter),
                     name = 'Year: Months') +
  labs(color = "Patient SARS-CoV2 Status") +
  scale_color_discrete(labels = c("Positive", "Negative/Unknown or Pre-COVID")) +
  theme_classic() + # theme_classic call
  theme(
    axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1),
    legend.title = element_text(size = rel(0.50)),
    legend.text  = element_text(size = rel(0.50)),
    legend.key.size = unit(0.50, "lines")
  )

#Now Add Annotations and Verticle Lines to Plot Based on What Data a Site Has Contributed
if (site_type == "Pre-COVID,COVID,Post-COVID") { # If Site Has Full Study Period (2018-2023)
  extended <- base_plot_covid +
    geom_vline(xintercept = 9, linetype = 2) +
    annotate("text", x = 5.667, y = 0.95, 
             label = 'Pre COVID-19\nPeriod', size = rel(1.5)) +
    geom_vline(xintercept = 18, linetype = 2) +
    annotate("text", x = 13.667, y = 0.95, 
             label = 'COVID-19\nPeriod', size = rel(1.5)) +
    annotate("text", x = 21, y = 0.95, 
             label = 'Post COVID-19\nPeriod', size = rel(1.5))
} else if (site_type == "COVID,Post-COVID") {
  extended <- base_plot +
    geom_vline(xintercept = 18, linetype = 2) +
    annotate("text", x = 13, y = 0.95, label = 'COVID-19\nPeriod', size = rel(1.5)) +
    annotate("text", x = 21, y = 0.95, label = 'Post COVID-19\nPeriod', size = rel(1.5))
} else if (site_type == "Pre-COVID,COVID") {
  extended <- base_plot +
    geom_vline(xintercept = 9, linetype = 2) +
    annotate("text", x = 5.667, y = 0.95, label = 'Pre COVID-19\nPeriod', size = rel(1.5)) +
    annotate("text", x = 13.667, y = 0.95, label = 'COVID-19\nPeriod',size = rel(1.5))
} else {
  extended <- base_plot
}
# Display the extended plot
extended
ggsave(paste0(site, '_proning_by_quarter_12_bycovid.pdf'),
       path = paste0(project_location, '/project_output/graphs/'))

```


```{r Proning and Patient Outcomes in Table Form}
to_tab <- c(
  'prone_12hour_outcome',
  'prone_24hour_outcome',
  'prone_72hour_outcome',
  'proned',
  'prone_episodes',
  'first_prone_episode_hours',
  'median_pt_prone_duration',
  'mean_pt_prone_duration',
  'time_to_prone_hrs',
  'in_hosp_death',
  'death_or_hospice'
)

#Factor Variables
factors_tab <- c(
   'prone_12hour_outcome',
   'prone_24hour_outcome',
   'prone_72hour_outcome',
  'proned',
  'in_hosp_death',
  'death_or_hospice'
)

#First Create Table with Ranges and Additional Statistics Including Numbers of Missing and Min/Max
tab2 <- CreateTableOne(vars = to_tab, data=prone_analytic_df, factorVars = factors_tab)
summary(tab2) #This will be in HTML File
#sink(paste0(project_location, '/project_output/', site, '_table2_missing_ranges.doc'))
#Repeat to Create Word File
#summary(tab2)
# Stop saving output to the file
#sink() 

#Now Create Table 2, One Stratified by Period One by Hospital and One by Period and COVID Status
tab2 <- CreateTableOne(vars = to_tab, strata="study_period", data=prone_analytic_df, factorVars = factors_tab, addOverall = TRUE) 
print(tab2)
tab2_excel <- as.data.frame(print(tab2, printToggle = FALSE, missing = TRUE)) |>
  mutate(site = site) |>
  relocate(site, .before = Overall) |>
  rownames_to_column('characteristic') |>
  rename(percent_missing = Missing) |>
  select(-test)
write.csv(tab2_excel, file=paste0(project_location, '/project_output/', site, '_Table_Prone_Outcomes_Period.csv'))
rm(tab2_excel)

#Now Create a Table 2, Stratified by COVID period
tab2 <- CreateTableOne(vars = to_tab, strata="period_sarscov2", data=prone_analytic_df, factorVars = factors_tab, addOverall = TRUE) 
print(tab2)
tab2_excel <- as.data.frame(print(tab2, printToggle = FALSE, missing = TRUE)) |>
  mutate(site = site) |>
  relocate(site, .before = Overall) |>
  rownames_to_column('characteristic') |>
  rename(percent_missing = Missing) |>
  select(-test)
write.csv(tab2_excel, file=paste0(project_location, '/project_output/', site, '_Table_Prone_Outcomes_by_Period_Sars-Cov2_Status.csv'))
rm(tab2_excel)

#Now Create a Table 2, Stratified by COVID period and SARS-Cov2 Status Only in COVID
tab2 <- CreateTableOne(vars = to_tab, strata="period_sarscov2_post", data=prone_analytic_df, factorVars = factors_tab, addOverall = TRUE) 
print(tab2)
tab2_excel <- as.data.frame(print(tab2, printToggle = FALSE, missing = TRUE)) |>
  mutate(site = site) |>
  relocate(site, .before = Overall) |>
  rownames_to_column('characteristic') |>
  rename(percent_missing = Missing) |>
  select(-test)
write.csv(tab2_excel, file=paste0(project_location, '/project_output/', site, '_Table_Prone_Outcomes_by_Period_Sars-Cov2_Status_Post.csv'))
rm(tab2_excel)

#Now Create a Table 2, Stratified by Hospital
if (n_hospitals>1) {
tab2 <- CreateTableOne(vars = to_tab, strata="hospital_id", data=prone_analytic_df, factorVars = factors_tab, addOverall = TRUE) 
print(tab2)
tab2_excel <- as.data.frame(print(tab2, printToggle = FALSE, missing = TRUE)) |>
  mutate(site = site) |>
  relocate(site, .before = Overall) |>
  rownames_to_column('characteristic') |>
  rename(percent_missing = Missing) |>
  select(-test)
} else {
tab2 <- CreateTableOne(vars = to_tab, data=prone_analytic_df, factorVars = factors_tab, addOverall = TRUE) 
print(tab2)
tab2_excel <- as.data.frame(print(tab2, printToggle = FALSE, missing = TRUE)) |>
  mutate(site = site) |>
  relocate(site, .before = Overall) |>
  rownames_to_column('characteristic') |>
  rename(percent_missing = Missing) 
}
write.csv(tab2_excel, file=paste0(project_location, '/project_output/', site, '_Table_Prone_Outcomes_by_Hospital.csv'))
rm(tab2_excel)


```

```{r Function to Summarise Logistic Regression Models}
model_summary_table <- function(model) {
 
#Uses Fixed Effects for Hospital in Sites with > 1 Hospital (and clustered standard errors)  
if (n_hospitals>1) {
  #Cluster Robust Standard Errors Using LM Test
  coef_table <- coeftest(model, 
                         vcovCL(model, cluster = model.frame(model)$hospital_id)) 
  # Cluster-robust covariance matrix
  clustered_se <- vcovCL(model, cluster = model.frame(model)$hospital_id)
  #Compute confidence intervals manually using the clustered SE
  conf_int <- coefci(model, vcov. = clustered_se)
}  else {
  coef_table <- coeftest(model)
  conf_int <- coefci(model)
}
 model_table <- data.frame(
  'site' = site,
  'n_observations' = nobs(model),
  'term' = rownames(coef_table),
  'estimate' = coef_table[, "Estimate"],
  'std_error' = coef_table[, "Std. Error"],
  'p_value' = coef_table[, "Pr(>|z|)"],
  'odds_ratio' = round(exp(coef_table[, "Estimate"]), 3),  # Exponentiated coefficients as Odds Ratios
  'lower_bound' = exp(conf_int[, 1]),  # Lower bound of CI from coefci()
  'upper_bound' = exp(conf_int[, 2])   # Upper bound of CI from coefci()
)
model_table <- model_table |>
  mutate(confidence_interval=paste0(round(lower_bound, 3), ' - ', round(upper_bound, 3))) |>
  mutate(n_hospitals=n_hospitals) |>
  relocate(confidence_interval, .before = lower_bound) |>
  relocate(n_hospitals, .after = n_observations) 
  rownames(model_table) <- NULL
return(model_table)
}

#Create a Function to Also Make Tables that Contain the Average Marginal Effects: THis Uses the maringaleffects::avg_slopes() function and using the'type==response' option yields estimates on the fraction scale (i.e, 0.1 is 10% absolute increase)
ame_summary <- function(model_list) {
  results <- list()  # Initialize an empty list to store results
  
  for (i in 1:length(model_list)) {
    # Extract the model formula as a single string
    formula_string <- paste(deparse(model_list[[i]]$formula), collapse = " ")
    
    # Compute and tidy the avg_slopes for the model
    slopes <- tidy(avg_slopes(model_list[[i]], type = "response"))
    
    # Add a column indicating the model index
    slopes <- slopes %>%
      mutate(model = paste0("Model ", i)) %>%
      mutate(across(c("estimate", "conf.low", "conf.high"), \(x) round(x, 4)))
    
    # Check if the model formula contains ':month_scaled' (interaction term)
    # Need to Also Check if this a Two Period Model (will have 'covid_period' in it)
    if (grepl("study_period:month_scaled", formula_string, ignore.case = TRUE)) {
      # Additional avg_slopes computation for interaction terms
      additional_slopes <- tidy(avg_slopes(model_list[[i]], variables = "month_scaled", by = "study_period")) %>%
        mutate(model = paste0("Model ", i, " - Interaction Terms")) %>%
        mutate(across(c("estimate", "conf.low", "conf.high"), \(x) round(x, 4)))
      
      # Combine the main slopes with additional slopes
      slopes <- bind_rows(slopes, additional_slopes)
    } else if (grepl("period_sarscov2_post:month_scaled", 
                     formula_string, ignore.case = TRUE)) {
      # Additional avg_slopes computation for interaction terms
      additional_slopes <- tidy(avg_slopes(model_list[[i]], variables = "month_scaled", by = "period_sarscov2_post")) %>%
        mutate(model = paste0("Model ", i, " - Interaction Terms")) %>%
        mutate(across(c("estimate", "conf.low", "conf.high"), \(x) round(x, 4)))
      
      # Combine the main slopes with additional slopes
      slopes <- bind_rows(slopes, additional_slopes)
    } else if (grepl("covid_period:month_scaled", formula_string, ignore.case = TRUE)) {
      
      # Additional avg_slopes computation for interaction terms
      additional_slopes <- tidy(avg_slopes(model_list[[i]], variables = "month_scaled", by = "covid_period")) %>%
        mutate(model = paste0("Model ", i, " - Interaction Terms")) %>%
        mutate(across(c("estimate", "conf.low", "conf.high"), \(x) round(x, 4)))
      
      # Combine the main slopes with additional slopes
      slopes <- bind_rows(slopes, additional_slopes)
    }
    
    # Add the result to the list
    results[[i]] <- slopes
    
    # Add an empty row to separate models (if not the last model)
    if (i < length(model_list)) {
      empty_row <- data.frame(matrix(NA, nrow = 1, ncol = ncol(slopes)))
      colnames(empty_row) <- colnames(slopes)
      empty_row$model <- NA  # Ensure model column exists
      results[[i]] <- bind_rows(results[[i]], empty_row)
    }
  }
  
  # Bind all results together into one table
  final_results <- bind_rows(results)
  
  return(final_results)  # Return the final combined table
}

#To Aid In Intrepretability Plan to Scale Each of The Continuous Variables
#Will Use Global Mean from Preliminary Data to Center Each Variable and then Scale by Reasonable Delta
prone_analytic_df <- prone_analytic_df |>
  mutate(
    age_scale=(age_at_admission-60)/10,
    bmi_scale=(bmi-30)/5,
    min_pf_ratio_scale=(min_pf_ratio-80)/10,
    sofa_score_scale=(sofa_score-9))
```


```{r Segmented Regression - Univariable}
#For Regressions Have the covid_sars_cov2+ period as reference
# Convert to an unordered factor
prone_analytic_df$period_sarscov2 <- factor(prone_analytic_df$period_sarscov2, ordered = FALSE)
# Now relevel to set 'COVID_SarsCov2Neg-Unk' as the reference group
prone_analytic_df$period_sarscov2 <- relevel(prone_analytic_df$period_sarscov2, ref = "COVID_SarsCov2Neg-Unk")
#Repeat for SarsCov2 only in the COVID Period
prone_analytic_df$period_sarscov2_post <- factor(prone_analytic_df$period_sarscov2_post, ordered = FALSE)
prone_analytic_df$period_sarscov2_post <- relevel(prone_analytic_df$period_sarscov2_post, ref = "COVID_SarsCov2Neg-Unk")


#For Sites with > 1 Hospital, Run Models with Fixed Effect for Hospital and Clustered Standard Errors 
#Uses Logistic Regression
#This is the Model Form for Evaluating the Step Changes NOT taking Into Account Time Trends

#First look at primary outcome proning within 12 hours by Study Period without SARS-COV2 Status. Then look at all proning. 
if (n_hospitals>1) {
model_form_step <- prone_12hour_outcome ~ study_period + factor(hospital_id)
unvar_mod <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
} else {
model_form_step <- prone_12hour_outcome ~ study_period
unvar_mod <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
}

summary(unvar_mod)
unvar_mod_summary <- model_summary_table(unvar_mod)

#First look at primary outcome proning within 12 hours by Study Period and COVID Status. Then look at all proning. 
if (n_hospitals>1) {
model_form_step <- prone_12hour_outcome ~ period_sarscov2 + factor(hospital_id)
unvar_mod_sarscov2 <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
} else {
model_form_step <- prone_12hour_outcome ~ period_sarscov2
unvar_mod_sarscov2 <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
}

summary(unvar_mod_sarscov2)
unvar_mod_summary_sarscov2_status <- model_summary_table(unvar_mod_sarscov2)

#Now primary outcome proning within 12 hours by Study Period and Sars Cov2 Status Broken Out for the COVID Period Only 
if (n_hospitals>1) {
model_form_step <- prone_12hour_outcome ~ period_sarscov2_post + factor(hospital_id)
unvar_mod_sarscov2_post <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
} else {
model_form_step <- prone_12hour_outcome ~ period_sarscov2_post
unvar_mod_sarscov2_post <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
}

summary(unvar_mod_sarscov2_post)
unvar_mod_summary_sarscov2_post <- model_summary_table(unvar_mod_sarscov2_post)

#This is the Model Form for Evaluating the Step Changes AND Time Trends
#Will Scale STudy Month So Coefficients Tell Us About Change in Log Odds Per 3 month Change
prone_analytic_df <- prone_analytic_df |>
  mutate(month_scaled=study_month/3)

if (n_hospitals>1) {
model_form_step <- prone_12hour_outcome ~ study_period + study_period:month_scaled + factor(hospital_id)
unvar_mod_interact <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
} else {
model_form_step <- prone_12hour_outcome ~ study_period + study_period:month_scaled 
unvar_mod_interact <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
}

summary(unvar_mod_interact)
unvar_modinteract_summary <- model_summary_table(unvar_mod_interact)

#This is the Model Form for Evaluating the Step Changes AND Time Trends With SARS Cov2 Status in the COVID Period
#Will Scale STudy Month So Coefficients Tell Us About Change in Log Odds Per 3 month Change
if (n_hospitals>1) {
  model_form_step <- prone_12hour_outcome ~ period_sarscov2_post + period_sarscov2_post:month_scaled + factor(hospital_id)
  unvar_mod_interact_sarscov2 <- glm(model_form_step,
                            data = prone_analytic_df, 
                            family=binomial)
} else {
  model_form_step <- prone_12hour_outcome ~ period_sarscov2_post + period_sarscov2_post:month_scaled
  unvar_mod_interact_sarscov2 <- glm(model_form_step,
                            data = prone_analytic_df, 
                            family=binomial)
}

summary(unvar_mod_interact_sarscov2)
unvar_modinteract_summary_sarscov2 <- model_summary_table(unvar_mod_interact_sarscov2)

#Summary Table With All Univariable (or site adjusted) Models
univar_table <- bind_rows(
  unvar_mod_summary,
  data.frame(site = c("", "Model With Sars_Cov2 Status")),
  unvar_mod_summary_sarscov2_status,
  data.frame(site = c("", "Model With Sars_Cov2 Status For COVID Period Only")),
  unvar_mod_summary_sarscov2_post,
  data.frame(site = c("", "Model With Interaction Terms - No SARS-Cov2 Status")),
  unvar_modinteract_summary,
  data.frame(site = c("", "Model With Interaction Terms - SARS-Cov2 Status During COVID")),
  unvar_modinteract_summary_sarscov2
)
# Write the table to a CSV file
write.csv(univar_table, paste0(project_location, '/project_output/', site, '_unadjusted_models.csv'))

#Output a Table of Average Marginal Effects Using the AME Function Created Above
model_list <- list(unvar_mod, unvar_mod_sarscov2, unvar_mod_sarscov2_post, 
                   unvar_mod_interact, unvar_mod_interact_sarscov2)
univar_ame <- ame_summary(model_list)
write.csv(univar_ame, paste0(project_location, '/project_output/', site, '_unadjusted_ame.csv'))
```


```{r Adjusted Segmented Regression}
#Agreed Upon Covariates (see DAG): Age, Gender, BMI, PF Ratio, Norepi Equivalents, sofa_score, hospital

#Full Model
model_form_step <- prone_12hour_outcome ~ study_period + age_scale + female + bmi_scale + factor(nee_pressor_dose) + sofa_score_scale +  
  min_pf_ratio_scale + factor(hospital_id)
#Now Modified for Each Site
if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
summary(multivar_mod)
multivar_mod_summary <- model_summary_table(multivar_mod)

#Full Model - with sars_cov2 stratification for all Periods
model_form_step <- prone_12hour_outcome ~ period_sarscov2 + age_scale + female + bmi_scale + factor(nee_pressor_dose) + sofa_score_scale +  
  min_pf_ratio_scale + factor(hospital_id)
#Now Modified for Each Site
if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod_sarscov2 <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
summary(multivar_mod_sarscov2)
multivar_mod_summary_sarscov2 <- model_summary_table(multivar_mod_sarscov2)

#Full Model - with sars_cov2 stratification for COVID Period ONly
model_form_step <- prone_12hour_outcome ~ period_sarscov2_post + age_scale + female + bmi_scale + factor(nee_pressor_dose) + sofa_score_scale +  
  min_pf_ratio_scale + factor(hospital_id)
#Now Modified for Each Site
if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod_sarscov2_post <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
summary(multivar_mod_sarscov2_post)
multivar_mod_summary_sarscov2_post <- model_summary_table(multivar_mod_sarscov2_post)

#This is the Model Form for Evaluating the Step Changes AND Time Trends
#Will Scale STudy Month So Coefficients Tell Us About Change in Log Odds Per 3 month Change
model_form_step <- prone_12hour_outcome ~ study_period + study_period:month_scaled + age_scale + female + bmi_scale + factor(nee_pressor_dose) + sofa_score_scale + min_pf_ratio_scale + factor(hospital_id)

if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod_interact <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
summary(multivar_mod_interact)
multivar_modinteract_summary <- model_summary_table(multivar_mod_interact)

#Evaluate Time Trends But With COVID Period Stratified by Sars-Cov2 Status
model_form_step <- prone_12hour_outcome ~ period_sarscov2_post +
              period_sarscov2_post:month_scaled + age_scale + female + 
  bmi_scale + factor(nee_pressor_dose) + sofa_score_scale + min_pf_ratio_scale + factor(hospital_id)

if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod_interact_sarscov2 <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
summary(multivar_mod_interact_sarscov2)
multivar_modinteract_summary_sarscov2 <- model_summary_table(multivar_mod_interact_sarscov2)

multivar_table <- bind_rows(
  multivar_mod_summary, 
  data.frame('site'=c('','Adjusted Model With Sars-CoV2 Status')),
  multivar_mod_summary_sarscov2, 
  data.frame('site'=c('','Adjusted Model With Sars-CoV2 Status For COVID ONly')),
  multivar_mod_summary_sarscov2_post, 
  data.frame('site'=c('','Adjusted Model With Interaction Term')),
  multivar_modinteract_summary,
  data.frame('site'=c('','Adjusted Model With Interaction Term - SarsCov2 Status')),
  multivar_modinteract_summary_sarscov2)
# Write the table to a CSV file
write.csv(multivar_table, paste0(project_location, '/project_output/', site, '_adjusted_models.csv'))

#Output a Table of Average Marginal Effects Using the AME Function Created Above
model_list <- list(multivar_mod, multivar_mod_sarscov2, 
                   multivar_mod_sarscov2_post, 
                   multivar_mod_interact, 
                   multivar_mod_interact_sarscov2)
multivar_ame <- ame_summary(model_list)
write.csv(multivar_ame, paste0(project_location, '/project_output/', site, '_adjusted_ame.csv'))
```


```{r Sensitivity Analysis #1 - 72 Hour Proning Outcome}
#Adjusted Models Only
#Agreed Upon Covariates (see DAG): Age, Gender, BMI, PF Ratio, Norepi Equivalents, sofa_score, hospital

#Full Model
model_form_step <- prone_72hour_outcome ~ study_period + age_scale + female + bmi_scale + factor(nee_pressor_dose) + sofa_score_scale +  
  min_pf_ratio_scale + factor(hospital_id)
#Now Modified for Each Site
if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
summary(multivar_mod)
multivar_mod_summary <- model_summary_table(multivar_mod)

#Full Model - with sars_cov2 stratification for all Periods
model_form_step <- prone_72hour_outcome ~ period_sarscov2 + age_scale + female + bmi_scale + factor(nee_pressor_dose) + sofa_score_scale +  
  min_pf_ratio_scale + factor(hospital_id)
#Now Modified for Each Site
if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod_sarscov2 <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
summary(multivar_mod_sarscov2)
multivar_mod_summary_sarscov2 <- model_summary_table(multivar_mod_sarscov2)

#Full Model - with sars_cov2 stratification for COVID Period ONly
model_form_step <- prone_72hour_outcome ~ period_sarscov2_post + age_scale + female + bmi_scale + factor(nee_pressor_dose) + sofa_score_scale +  
  min_pf_ratio_scale + factor(hospital_id)
#Now Modified for Each Site
if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod_sarscov2_post <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
summary(multivar_mod_sarscov2_post)
multivar_mod_summary_sarscov2_post <- model_summary_table(multivar_mod_sarscov2_post)

#This is the Model Form for Evaluating the Step Changes AND Time Trends
#Will Scale STudy Month So Coefficients Tell Us About Change in Log Odds Per 3 month Change
model_form_step <- prone_72hour_outcome ~ study_period + study_period:month_scaled + age_scale + female + bmi_scale + factor(nee_pressor_dose) + sofa_score_scale + min_pf_ratio_scale + factor(hospital_id)

if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod_interact <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
summary(multivar_mod_interact)
multivar_modinteract_summary <- model_summary_table(multivar_mod_interact)

#Evaluate Time Trends But With COVID Period Stratified by Sars-Cov2 Status
model_form_step <- prone_72hour_outcome ~ period_sarscov2_post +
              period_sarscov2_post:month_scaled + age_scale + female + 
  bmi_scale + factor(nee_pressor_dose) + sofa_score_scale + min_pf_ratio_scale + factor(hospital_id)

if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod_interact_sarscov2 <- glm(model_form_step,
          data = prone_analytic_df, 
          family=binomial)
summary(multivar_mod_interact_sarscov2)
multivar_modinteract_summary_sarscov2 <- model_summary_table(multivar_mod_interact_sarscov2)

multivar_table <- bind_rows(
  multivar_mod_summary, 
  data.frame('site'=c('','Adjusted Model With Sars-CoV2 Status')),
  multivar_mod_summary_sarscov2, 
  data.frame('site'=c('','Adjusted Model With Sars-CoV2 Status For COVID ONly')),
  multivar_mod_summary_sarscov2_post, 
  data.frame('site'=c('','Adjusted Model With Interaction Term')),
  multivar_modinteract_summary,
  data.frame('site'=c('','Adjusted Model With Interaction Term - SarsCov2 Status')),
  multivar_modinteract_summary_sarscov2)
# Write the table to a CSV file
write.csv(multivar_table, paste0(project_location, '/project_output/', site, '_adjusted_models_prone72.csv'))

#Output a Table of Average Marginal Effects Using the AME Function Created Above
model_list <- list(multivar_mod, multivar_mod_sarscov2, 
                   multivar_mod_sarscov2_post, 
                   multivar_mod_interact, 
                   multivar_mod_interact_sarscov2)
multivar_ame <- ame_summary(model_list)
write.csv(multivar_ame, paste0(project_location, '/project_output/', site, '_adjusted_ame_prone72.csv'))
```

```{r Sensitivity Analysis #2 - Severe AHRF Only}
#Adjusted Analysis - Subset to Severe ARDS

#Full Model
model_form_step <- prone_12hour_outcome ~ study_period + age_scale + female + bmi_scale + factor(nee_pressor_dose) + sofa_score_scale +  
  min_pf_ratio_scale + factor(hospital_id)
#Now Modified for Each Site
if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod <- glm(model_form_step,
          data = subset(prone_analytic_df, severe_ards==1), 
          family=binomial)
summary(multivar_mod)
multivar_mod_summary <- model_summary_table(multivar_mod)

#Full Model - with sars_cov2 stratification for all Periods
model_form_step <- prone_12hour_outcome ~ period_sarscov2 + age_scale + female + bmi_scale + factor(nee_pressor_dose) + sofa_score_scale +  
  min_pf_ratio_scale + factor(hospital_id)
#Now Modified for Each Site
if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod_sarscov2 <- glm(model_form_step,
          data = subset(prone_analytic_df, severe_ards==1), 
          family=binomial)
summary(multivar_mod_sarscov2)
multivar_mod_summary_sarscov2 <- model_summary_table(multivar_mod_sarscov2)

#Full Model - with sars_cov2 stratification for COVID Period ONly
model_form_step <- prone_12hour_outcome ~ period_sarscov2_post + age_scale + female + bmi_scale + factor(nee_pressor_dose) + sofa_score_scale +  
  min_pf_ratio_scale + factor(hospital_id)
#Now Modified for Each Site
if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod_sarscov2_post <- glm(model_form_step,
          data = subset(prone_analytic_df, severe_ards==1), 
          family=binomial)
summary(multivar_mod_sarscov2_post)
multivar_mod_summary_sarscov2_post <- model_summary_table(multivar_mod_sarscov2_post)

#This is the Model Form for Evaluating the Step Changes AND Time Trends
#Will Scale STudy Month So Coefficients Tell Us About Change in Log Odds Per 3 month Change
model_form_step <- prone_12hour_outcome ~ study_period + study_period:month_scaled + age_scale + female + bmi_scale + factor(nee_pressor_dose) + sofa_score_scale + min_pf_ratio_scale + factor(hospital_id)

if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod_interact <- glm(model_form_step,
          data = subset(prone_analytic_df, severe_ards==1), 
          family=binomial)
summary(multivar_mod_interact)
multivar_modinteract_summary <- model_summary_table(multivar_mod_interact)

#Evaluate Time Trends But With COVID Period Stratified by Sars-Cov2 Status
model_form_step <- prone_12hour_outcome ~ period_sarscov2_post +
              period_sarscov2_post:month_scaled + age_scale + female + 
  bmi_scale + factor(nee_pressor_dose) + sofa_score_scale + min_pf_ratio_scale + factor(hospital_id)

if (n_hospitals > 1) {
  # Keep the formula as is
  model_form_step <- model_form_step
} else if (n_hospitals == 1) {
  # Remove 'factor(hospital_id)'
  model_form_step <- update(model_form_step, . ~ . - factor(hospital_id))
}

multivar_mod_interact_sarscov2 <- glm(model_form_step,
          data = subset(prone_analytic_df, severe_ards==1), 
          family=binomial)
summary(multivar_mod_interact_sarscov2)
multivar_modinteract_summary_sarscov2 <- model_summary_table(multivar_mod_interact_sarscov2)

multivar_table <- bind_rows(
  multivar_mod_summary, 
  data.frame('site'=c('','Adjusted Model With Sars-CoV2 Status')),
  multivar_mod_summary_sarscov2, 
  data.frame('site'=c('','Adjusted Model With Sars-CoV2 Status For COVID ONly')),
  multivar_mod_summary_sarscov2_post, 
  data.frame('site'=c('','Adjusted Model With Interaction Term')),
  multivar_modinteract_summary,
  data.frame('site'=c('','Adjusted Model With Interaction Term - SarsCov2 Status')),
  multivar_modinteract_summary_sarscov2)
# Write the table to a CSV file
write.csv(multivar_table, paste0(project_location, '/project_output/', site, '_adjusted_models_severe_ards.csv'))

#Output a Table of Average Marginal Effects Using the AME Function Created Above
model_list <- list(multivar_mod, multivar_mod_sarscov2, 
                   multivar_mod_sarscov2_post, 
                   multivar_mod_interact, 
                   multivar_mod_interact_sarscov2)
multivar_ame <- ame_summary(model_list)
write.csv(multivar_ame, paste0(project_location, '/project_output/', site, '_adjusted_ame_severe_ards.csv'))

```


```{r Last Piece Create a Table with Observed and Expeted Proning Counts/Rates by Period and Hospital}
#This Aggregate Data Will be Used for Risk and Reliability Adjusted Proning Rate Estimates by Hospital and Period
#Generate Prone Propensity Score (Don't Account for Hospital or Study Period Here)
alt_form <- prone_12hour_outcome ~ age_scale  + female +
  bmi + factor(nee_pressor_dose) +sofa_score_scale + min_pf_ratio_scale

alt_mod <- glm(alt_form,
               data = prone_analytic_df, 
               family=binomial)

#Collect Coefficient and Covariance Matrix for Centralized Creation of Global Coefficient
propensity_build_estimates <- model_summary_table(alt_mod)
propensity_build_covariance <- as.data.frame(vcov(alt_mod))
write.csv(propensity_build_estimates, paste0(project_location, '/project_output/', site, '_propensity_build_estimates.csv'))
write.csv(propensity_build_covariance, paste0(project_location, '/project_output/', site, '_propensity_build_covariance.csv'))

#Extract Dataset from Model
df <- alt_mod$data
#Propensity Score Taken from Prediction
df$prone_propensity <- predict(alt_mod, newdata=df, type = 'response')

#Each SIte Aggregates Data
#Create Aggregate Dataset with Observed and Expected Events as well as N
df_agg <- df |>
  group_by(hospital_id, study_period) |>
  summarise(
    n_patients = n(),
    observed_prone = sum(prone_12hour_outcome, na.rm=TRUE),
    prone_rate_observed = mean(prone_12hour_outcome, na.rm=TRUE),
    prone_rate_adjust = mean(prone_propensity, na.rm = TRUE)
  ) |>
  ungroup() |>
  filter(!is.na(hospital_id)) |>
  mutate(expected_prone=round((prone_rate_adjust*n_patients), digits = 0)) |>
  mutate(not_prone=n_patients-observed_prone)

#Save CSV
write.csv(df_agg, paste0(project_location, '/project_output/', site, '_aggregate_expected_prone.csv'))

#Create Aggregate Dataset with Observed and Expected Events as well as N Broken Out by SARS Cov2 Status
df_agg_sars_status <- df |>
  group_by(hospital_id, period_sarscov2) |>
  summarise(
    n_patients = n(),
    observed_prone = sum(prone_12hour_outcome, na.rm=TRUE),
    prone_rate_observed = mean(prone_12hour_outcome, na.rm=TRUE),
    prone_rate_adjust = mean(prone_propensity, na.rm = TRUE)
  ) |>
  ungroup() |>
  filter(!is.na(hospital_id)) |>
  mutate(expected_prone=round((prone_rate_adjust*n_patients), digits = 0)) |>
  mutate(not_prone=n_patients-observed_prone)

#Save CSV
write.csv(df_agg_sars_status, paste0(project_location, '/project_output/', site, '_aggregate_expected_prone_wCOVID-Status.csv'))

#Create Aggregate Dataset with Observed and Expected Events as well as N Broken Out by SARS Cov2 Status
df_agg_sars_status <- df |>
  group_by(hospital_id, period_sarscov2_post) |>
  summarise(
    n_patients = n(),
    observed_prone = sum(prone_12hour_outcome, na.rm=TRUE),
    prone_rate_observed = mean(prone_12hour_outcome, na.rm=TRUE),
    prone_rate_adjust = mean(prone_propensity, na.rm = TRUE)
  ) |>
  ungroup() |>
  filter(!is.na(hospital_id)) |>
  mutate(expected_prone=round((prone_rate_adjust*n_patients), digits = 0)) |>
  mutate(not_prone=n_patients-observed_prone)

#Save CSV
write.csv(df_agg_sars_status, paste0(project_location, '/project_output/', site, '_aggregate_expected_prone_wCOVID-Status_post.csv'))
```

```{r}
cat('This code is DONE!!! Thank you for your support for this CLIF project!')
```
